+++ R/ab.R
@@ -28,7 +28,9 @@
 #' @inheritSection WHOCC WHOCC
 #' @export
 #' @importFrom dplyr %>% filter slice pull
-#' @details Use the \code{\link{ab_property}} functions to get properties based on the returned ATC code, see Examples.
+#' @details All entries in the \code{\{link{antibiotics}} data set have three different identifiers: a human readable EARS-Net code (column \code{ab}, used by ECDC and WHONET), an ATC code (column \code{atc}, used by WHO), and a CID code (column \code{cid}, Compound ID, used by PubChem). The data set contains more than 5,000 official brand names from many different countries, as found in PubChem.
+#'
+#' Use the \code{\link{ab_property}} functions to get properties based on the returned ATC code, see Examples.
 #'
 #' In the ATC classification system, the active substances are classified in a hierarchy with five different levels. The system has fourteen main anatomical/pharmacological groups or 1st levels. Each ATC main group is divided into 2nd levels which could be either pharmacological or therapeutic groups.  The 3rd and 4th levels are chemical, pharmacological or therapeutic subgroups and the 5th level is the chemical substance.  The 2nd, 3rd and 4th levels are often used to identify pharmacological subgroups when that is considered more appropriate than therapeutic or chemical subgroups.
 #'   Source: \url{https://www.whocc.no/atc/structure_and_principles/}
@@ -38,7 +40,7 @@
 #' WHONET 2019 software: \url{http://www.whonet.org/software.html}
 #'
 #' European Commission Public Health PHARMACEUTICALS - COMMUNITY REGISTER: \url{http://ec.europa.eu/health/documents/community-register/html/atc.htm}
-#' @return Character (vector) with class \code{"act"}. Unknown values will return \code{NA}.
+#' @return Character (vector) with class \code{"ab"}. Unknown values will return \code{NA}.
 #' @seealso \code{\link{antibiotics}} for the dataframe that is being used to determine ATCs.
 #' @inheritSection AMR Read more on our website!
 #' @examples
@@ -64,9 +66,9 @@ as.ab <- function(x) {
   }
   x_bak <- x
   # remove suffices
-  x_bak_clean <- gsub("_(mic|rsi)$", "", x)
+  x_bak_clean <- gsub("_(mic|rsi|disk|disc)$", "", x)
   # remove disk concentrations, like LVX_NM -> LVX
-  x_bak_clean <- gsub("_[A-Z]{2}[0-9_]{0,3}$", "", x_bak_clean)
+  x_bak_clean <- gsub("_[A-Z]{2}[0-9_]{0,3}$", "", x_bak_clean, ignore.case = TRUE)
   # clean rest of it
   x_bak_clean <- gsub("[^a-zA-Z0-9/-]", "", x_bak_clean)
   # keep only a-z when it's not an ATC code or only numbers