+++ tests/testthat/test-count.R
@@ -51,14 +51,19 @@ test_that("counts work", {
   # count_df
   expect_equal(
     septic_patients %>% select(AMX) %>% count_df() %>% pull(Value),
-    c(septic_patients$AMX %>% count_S(),
-      septic_patients$AMX %>% count_I(),
+    c(septic_patients$AMX %>% count_SI(),
       septic_patients$AMX %>% count_R())
   )
   expect_equal(
     septic_patients %>% select(AMX) %>% count_df(combine_IR = TRUE) %>% pull(Value),
     c(septic_patients$AMX %>% count_S(),
       septic_patients$AMX %>% count_IR())
+  )
+  expect_equal(
+    septic_patients %>% select(AMX) %>% count_df(combine_SI = FALSE) %>% pull(Value),
+    c(septic_patients$AMX %>% count_S(),
+      septic_patients$AMX %>% count_I(),
+      septic_patients$AMX %>% count_R())
   )
 
   # warning for speed loss